Replace all floats with double

+++ Positioning/LatitudeDegree.cpp
@@ -5,11 +5,11 @@ using namespace std;
 
 LatitudeDegree::LatitudeDegree(const std::string &koordString, char direction) {
     int deg = 0;
-    float min = 0;
+    double min = 0;
     stringstream(koordString.substr(0, 2)) >> deg;
     stringstream(koordString.substr(2, 10)) >> min;
 
-    float jdeg = deg + (min / 60);
+    double jdeg = deg + (min / 60);
     if (direction == 'S') {
         jdeg = -jdeg;
     }
@@ -22,12 +22,12 @@ LatitudeDegree operator-(const LatitudeDegree &dg1, const LatitudeDegree &dg2) {
 }
 
 //Returns Distance to Point Zero
-float LatitudeDegree::toMeters() const {
+double LatitudeDegree::toMeters() const {
     return std::abs(degrees) * 60 * 1852.216;
 }
 
 
-LatitudeDegree::LatitudeDegree(float degrees) {
+LatitudeDegree::LatitudeDegree(double degrees) {
     if (degrees < -90 || degrees > 90) {
         return;
     }
