+++ Positioning/Position.cpp
@@ -7,12 +7,12 @@
  * @param p the Positon which is compared with this pos
  * @return the distance from Position to p in meters
  */
-float Position::distanceTo(const Position &p) const {
+double Position::distanceTo(const Position &p) const {
 
     LatitudeDegree dlat = this->latitude - p.latitude;
     LongitudeDegree dlong = this->longitude - p.longitude;
 
-    float avgDeg = (latitude.getDegrees() + p.latitude.getDegrees()) / 2;
+    double avgDeg = (latitude.getDegrees() + p.latitude.getDegrees()) / 2;
     return sqrt(dlat.toMeters() * dlat.toMeters() + dlong.toMeters(avgDeg) * dlong.toMeters(avgDeg));
 }
 
@@ -22,8 +22,8 @@ void Position::setPosition(LatitudeDegree &x, LongitudeDegree &y) {
 }
 
 Position getMedian(const vector<Position> &p) {
-    float latDeg = 0;
-    float lonDeg = 0;
+    double latDeg = 0;
+    double lonDeg = 0;
 
     for (const auto &i : p) {
         latDeg += i.latitude.getDegrees();
@@ -38,7 +38,7 @@ Position getMedian(const vector<Position> &p) {
 string Position::toString() const {
     string output;
 
-    double lat = (double) latitude.getDegrees();
+    double lat = latitude.getDegrees();
     unsigned short latDegree, latMinutes, latSeconds;
     char latDir = (lat > 0) ? 'N' : 'S';
     //lat = std::abs(lat);     //Ruft den Fehler: 50,00, 0, hervor
@@ -53,9 +53,9 @@ string Position::toString() const {
     latSeconds = (unsigned short) (lat + 0.500001);
 
     sprintf(buff,"lat: %2.0d,%2.0d,%2.0d %c \r\n", latDegree, latMinutes, latSeconds, latDir);  
-    output.assign(buff);    
+    output.assign(buff);
 
-    double lon = (double) longitude.getDegrees();
+    double lon = longitude.getDegrees();
     unsigned short lonDegree, lonMinutes, lonSeconds;
     char lonDir = (lon > 0) ? 'E' : 'W';
     //lon = std::abs(lon);
